Terminando

+++ src/main/java/berral/francisco/Film_Manager/Proyection_Menu_Controller.java
@@ -12,6 +12,7 @@ import berral.francisco.Film_Manager.model.DAO.ProyectionDAO;
 import berral.francisco.Film_Manager.model.DataObject.Cinema;
 import berral.francisco.Film_Manager.model.DataObject.Production;
 import berral.francisco.Film_Manager.model.DataObject.Proyection;
+import berral.francisco.Film_Manager.utils.Log;
 import berral.francisco.Film_Manager.utils.Message;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
@@ -25,7 +26,13 @@ import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.input.MouseEvent;
 
+/**
+ * Controlador "Proyection_Menu_Controller" que implementa "Initializable"
+ * @author Francisco José Berral Zafra
+ *
+ */
 public class Proyection_Menu_Controller implements Initializable{
 	
 	CinemaDAO cDAO = new CinemaDAO();
@@ -59,11 +66,19 @@ public class Proyection_Menu_Controller implements Initializable{
 	@FXML 
 	private TableColumn<Proyection, LocalDate> finishCol;
 	
+	/**
+	 * Cambia a la Vista "Main"
+	 * @throws IOException Lanza excepción en caso de error
+	 */
 	@FXML
 	private void switchToMain() throws IOException {
 		App.setRoot("main_menu_view");
 	}
 	
+	/**
+	 * Método asignado al botón "ADD" que permite añadir una proyección
+	 * @throws IOException Lanza excepción en caso de error
+	 */
 	@FXML
 	private void addProyection() throws IOException {
 		Cinema c = comCinema.getSelectionModel().getSelectedItem(); 
@@ -71,28 +86,34 @@ public class Proyection_Menu_Controller implements Initializable{
 		LocalDate sD = dateStart.getValue();
 		LocalDate fD = dateFinish.getValue();
 		
-		Proyection pr = pDAO.get(c.getID_C(), p.getID_F(), sD);
+		if(comCinema.getSelectionModel().getSelectedItem() != null && comProd.getSelectionModel().getSelectedItem() != null && dateStart.getValue() != null && dateFinish.getValue() != null) {
+			Proyection pr = pDAO.get(c.getID_C(), p.getID_F(), sD);
 		
-		if(pr == null) {
-			if(comCinema.getSelectionModel().getSelectedItem() != null && comProd.getSelectionModel().getSelectedItem() != null && dateStart.getValue() != null && dateFinish.getValue() != null) {
+			if(pr == null) {
 				if(fD.isAfter(sD)) {
 					Proyection pro = new Proyection(c, p, sD, fD);
 					pDAO.insert(pro);
-					dateStart.getEditor().clear();
-					dateFinish.getEditor().clear();
 					initialize(null, null);
 					Message.alert("SUCCESS", "OPERATION SUCCESSFULLY", "PROYECTION HAS BEEN ADDED");
+					Log.info("PROYECTION HAS BEEN ADDED");
 				}else {
 					Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "THE END DATE MUST BE LATER THAN THE START DATE");
+					Log.severe("THE END DATE MUST BE LATER THAN THE START DATE");
 				}
 			}else {
-				Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "ALL FIELDS ARE REQUIRED");
+				Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "PROYECTION ALREADY EXISTS");
+				Log.severe("PROYECTION ALREADY EXISTS");
 			}
 		}else {
-			Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "PROYECTION ALREADY EXISTS");
+			Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "ALL FIELDS ARE REQUIRED");
+			Log.severe("ALL FIELDS ARE REQUIRED");
 		}
 	}
 	
+	/**
+	 * Método asignado al botón "DELETE" que permite eliminar una proyección
+	 * @throws IOException Lanza excepción en caso de error
+	 */
 	@FXML
 	private void deleteProyection() throws IOException {
 		Cinema c = comCinema.getSelectionModel().getSelectedItem(); 
@@ -101,16 +122,25 @@ public class Proyection_Menu_Controller implements Initializable{
 		
 		if(comCinema.getSelectionModel().getSelectedItem() != null && comProd.getSelectionModel().getSelectedItem() != null && dateStart.getValue() != null) {
 			Proyection pr = pDAO.get(c.getID_C(), p.getID_F(), sD);
-			pDAO.delete(pr);
-			dateStart.getEditor().clear();
-			dateFinish.getEditor().clear();
-			initialize(null, null);
-			Message.alert("SUCCESS", "OPERATION SUCCESSFULLY", "PROYECTION HAS BEEN DELETED");
+			if(pr != null) {
+				pDAO.delete(pr);
+				initialize(null, null);
+				Message.alert("SUCCESS", "OPERATION SUCCESSFULLY", "PROYECTION HAS BEEN DELETED");
+				Log.info("PROYECTION HAS BEEN DELETED");
+			}else {
+				Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "PROYECTION NOT FOUND");
+				Log.severe("PROYECTION NOT FOUND");
+			}
 		}else {
-			Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "FIELDS ID_C, ID_F AND STARTDATE FROM PROYECTION ARE REQUIRED");	
+			Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "FIELDS ID_C, ID_F AND STARTDATE FROM PROYECTION ARE REQUIRED");
+			Log.severe("FIELDS ID_C, ID_F AND STARTDATE FROM PROYECTION ARE REQUIRED");
 		}
 	}
 
+	/**
+	 * Método asignado al botón "UPDATE" que permite actualizar una proyección
+	 * @throws IOException Lanza excepción en caso de error
+	 */
 	@FXML
 	private void modifyProyection() throws IOException {
 		Cinema c = comCinema.getSelectionModel().getSelectedItem(); 
@@ -121,21 +151,45 @@ public class Proyection_Menu_Controller implements Initializable{
 		if(comCinema.getSelectionModel().getSelectedItem() != null && comProd.getSelectionModel().getSelectedItem() != null && dateStart.getValue() != null) {
 			Proyection pr = pDAO.get(c.getID_C(), p.getID_F(), sD);
 			if(pr != null) {
-				pr.setC(c);
-				pr.setP(p);
-				pr.setStartDate(sD);
-				pr.setFinishDate(fD);
-				pDAO.update(pr);
-				dateStart.getEditor().clear();
-				dateFinish.getEditor().clear();
-				initialize(null, null);
-				Message.alert("SUCCESS", "OPERATION SUCCESSFULLY", "PROYECTION HAS BEEN MODIFIED");
+				if(fD.isAfter(sD)) {
+					pr.setC(c);
+					pr.setP(p);
+					pr.setStartDate(sD);
+					pr.setFinishDate(fD);
+					pDAO.update(pr);
+					initialize(null, null);
+					Message.alert("SUCCESS", "OPERATION SUCCESSFULLY", "PROYECTION HAS BEEN MODIFIED");
+					Log.info("PROYECTION HAS BEEN MODIFIED");
+				}else {
+					Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "THE END DATE MUST BE LATER THAN THE START DATE");
+					Log.severe("THE END DATE MUST BE LATER THAN THE START DATE");
+				}
+			}else {
+				Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "PROYECTION NOT FOUND");
+				Log.severe("PROYECTION NOT FOUND");
 			}
 		}else {
-			Message.error("ERROR", "ERROR WHEN ENTERING CINEMA", "FIELDS ID_C, ID_F AND STARTDATE FROM PROYECTION ARE REQUIRED");	
+			Message.error("ERROR", "ERROR WHEN ENTERING PROYECTION", "FIELDS ID_C, ID_F AND STARTDATE FROM PROYECTION ARE REQUIRED");	
+			Log.severe("FIELDS ID_C, ID_F AND STARTDATE FROM PROYECTION ARE REQUIRED");
+		}
+	}
+	
+	@FXML
+	private void onEdit() {
+		if(proyectionTable.getSelectionModel().getSelectedItem() != null) {
+			Proyection p = proyectionTable.getSelectionModel().getSelectedItem();
+			Cinema c = cDAO.get(p.getC().getID_C());
+			comCinema.setValue(c);
+			Production pr = ProductionDAO.get(p.getP().getID_F());
+			comProd.setValue(pr);
+			dateStart.setValue(p.getStartDate());
+			dateFinish.setValue(p.getFinishDate());
 		}
 	}
 
+	/**
+	 * Método que permite setear en las celdas de la tabla los valores de los ComboBoxs y de los DatePickers
+	 */
 	@Override
 	public void initialize(URL location, ResourceBundle resources) {
 		List<Cinema> cinemaList = (List<Cinema>) cDAO.getAll();
@@ -171,5 +225,11 @@ public class Proyection_Menu_Controller implements Initializable{
 		});
 		
 		proyectionTable.setItems(FXCollections.observableArrayList(obsPro));
+		
+		proyectionTable.setOnMouseClicked((MouseEvent event) -> {
+			if (event.getClickCount() > 1) {
+				onEdit();
+			}
+		});
 	}
 }
